Improved New_Find_X function
Improved the New_Find_X function by allowing the input of the equation as an argument, allowing it to be used by other programs / files. Everything inside the function is now isolated, now not needing anything outside of it to function. Imposing the second function might be a challenge for this reason. test_stuff tests the function, with success!